refactor: Generate background service worker from typescript
