+++ package.json
@@ -6,6 +6,7 @@
     "ng": "ng",
     "start": "ng serve",
     "build": "ng build --configuration production --output-hashing none",
+    "postbuild": "webpack",
     "test": "jest",
     "test:watch": "jest --watch",
     "test:coverage": "npm run test -- --coverage=true",
@@ -62,7 +63,10 @@
     "ng-packagr": "^13.3.1",
     "prettier": "^2.2.1",
     "protractor": "~7.0.0",
+    "ts-loader": "^9.3.0",
     "ts-node": "^10.7.0",
-    "typescript": "~4.6.3"
+    "typescript": "~4.6.3",
+    "webpack": "^5.72.0",
+    "webpack-cli": "^4.9.2"
   }
 }
