+++ src/app/utils/index.ts
@@ -1,2 +1,3 @@
-export * from './tab';
 export * from './browser';
+export * from './models';
+export * from './tab';
