+++ src/app/utils/models.ts
@@ -1,0 +1,33 @@
+/**
+ * Storage key used to store tab groups in local storage.
+ */
+export const storageKey = 'tabs';
+
+/**
+ * Matches domain name.
+ */
+export const domainRegExp = new RegExp('([^.]*).([^.]*)$');
+
+/**
+ * Tab type.
+ */
+export type BrowserTab = Pick<chrome.tabs.Tab, 'id' | 'url' | 'favIconUrl' | 'title'>;
+
+/**
+ * Unique domains in tab group.
+ */
+export interface Domain {
+  name: string;
+  icon: string;
+  count: number;
+}
+
+/**
+ * Group that contains tabs and is saved to local storage.
+ */
+export interface TabGroup {
+  id: string;
+  timestamp: number;
+  tabs: BrowserTab[];
+  domains: Domain[];
+}
