+++ src/app/utils/tab.ts
@@ -1,39 +1,6 @@
 import { saveAs } from 'file-saver';
 import selectFiles from 'select-files';
-
-/**
- * Storage key used to store tab groups in local storage.
- */
-export const storageKey = 'tabs';
-
-/**
- * Matches domain name.
- */
-export const domainRegExp = new RegExp('([^.]*).([^.]*)$');
-
-/**
- * Tab type.
- */
-export type BrowserTab = Pick<chrome.tabs.Tab, 'id' | 'url' | 'favIconUrl' | 'title'>;
-
-/**
- * Unique domains in tab group.
- */
-export interface Domain {
-  name: string;
-  icon: string;
-  count: number;
-}
-
-/**
- * Group that contains tabs and is saved to local storage.
- */
-export interface TabGroup {
-  id: string;
-  timestamp: number;
-  tabs: BrowserTab[];
-  domains: Domain[];
-}
+import { BrowserTab, Domain, storageKey, TabGroup } from './models';
 
 /**
  * Saves specified tab groups to local storage.
